+++ src/Code_LayerHold.h
@@ -3,18 +3,18 @@
 
 #include <inttypes.h>
 #include <Code.h>
-#include "StateLayers.h"
+#include "LayerState.h"
 
-/* Code_LayerHold calls StateLayers when pressed to change activeLayer.
+/* Code_LayerHold calls LayerState when pressed to change activeLayer.
 */
 class Code_LayerHold : public Code
 {
     private:
         const uint8_t layer;
-        StateLayers& refStateLayers;
+        LayerState& refLayerState;
     public:
-        Code_LayerHold(const uint8_t layer, StateLayers& refStateLayers)
-            : layer(layer), refStateLayers(refStateLayers) {}
+        Code_LayerHold(const uint8_t layer, LayerState& refLayerState)
+            : layer(layer), refLayerState(refLayerState) {}
         virtual void press();
         virtual void release();
 };