+++ src/Code_LayeredCodeScBase.h
@@ -7,7 +7,7 @@
 /* Class Code_LayeredCodeScBase is a 2-layer code, one object for each layer e.g.
     layer0: ms_up    //mouse up
     layer1: KEY_UP   //up arrow
-When the key is pressed, the active layer is retrieved from refStateLayers,
+When the key is pressed, the active layer is retrieved from refLayerState,
  and the object for the active layer is sent to USB.
 */
 class Code_LayeredCodeScBase : public Code