+++ src/Code_LayeredScSc.h
@@ -2,18 +2,18 @@
 #define CODE_LAYEREDSCSC_H
 #include <Arduino.h>
 #include <inttypes.h>
-#include <StateLayersInterface.h>
+#include <LayerStateInterface.h>
 #include <Code_LayeredScScBase.h>
 
 /* Class Code_LayeredScSc is composed of two scancodes; "S" stands for Scancode.
-layer is retreived from refStateLayers.
+layer is retreived from refLayerState.
     when layer=0, press sends scancode0
     when layer=1, press sends scancode1
 */
 class Code_LayeredScSc : public Code_LayeredScScBase
 {
     private:
-        static StateLayersInterface& refStateLayers;
+        static LayerStateInterface& refLayerState;
     public:
         Code_LayeredScSc(const uint16_t scancode0, const uint16_t scancode1)
             : Code_LayeredScScBase(scancode0, scancode1) { }