+++ src/LayerState.cpp
@@ -1,11 +1,11 @@
-#include "StateLayers.h"
+#include "LayerState.h"
 
-void StateLayers::hold(const uint8_t layer)
+void LayerState::hold(const uint8_t layer)
 {
     setActiveLayer(layer);
 }
 
-void StateLayers::unhold(const uint8_t layer)
+void LayerState::unhold(const uint8_t layer)
 {
     if (layer == activeLayer);
     {
@@ -13,19 +13,19 @@ void StateLayers::unhold(const uint8_t layer)
     }
 }
 
-void StateLayers::lock(const uint8_t layer)
+void LayerState::lock(const uint8_t layer)
 {
     setActiveLayer(layer);
     lockedLayer = layer;
 }
 
 //Derived classes override setActiveLayer() to also set LED indicator lights.
-void StateLayers::setActiveLayer(const uint8_t layer)
+void LayerState::setActiveLayer(const uint8_t layer)
 {
     activeLayer = layer;
 }
 
-uint8_t StateLayers::getActiveLayer()
+uint8_t LayerState::getActiveLayer()
 {
     return activeLayer;
 }